Split Prometheus Data Source and Scrape
Prometheus has two roles when used with Grafana. The first is to scrape
the metrics and the second is to provide a data source to grafana.

The problem we keep having is we peak on CPU and Memory, and this
happens when we run grafana to query data. During this though the
prometheus instance is impacted and we can have problems with the
scrapes.

So the solution is to use two prometheus instances, one exclusivly for
scrapes and the other as a grafana data source.