Use CF instance label in internal apps
By default the instance label is set to the PaaS internal IP
address/port. It is not useful to us and it creates a lot of time series
in prometheus and influxdb as it's different everytime the app instance
moves to another VM.

Now if the target provides a "app_instance" label, we replace the
default instance label and drop app_instance.

If app_instance is not present, nothing changes.