Use CF instance label from the paas exporter
The CF app instance number is provided by the paas prometheus exporter
but it conflicts with the prometheus own instance label, which we're
not interested in.

By using "honor_labels" we tell prometheus to replace its own label with
the one from the exporter.

This applies to all the exporters but only the paas one sends an
instance label so other ones are not impacted.