Rename the postgres service key
A service key is created to allow the postgres_prometheus_exporter to
connect to each app's postgres service.  This key is named after the
app's postgres service which conflicts with a key created in the app's
terraform file.  The key has been renamed to make it clear it belongs
to the postgres_prometheus_exporter