Deprecated input & add optional Slack receivers
DEPRECATED INPUT: alertmanager_slack_channel, this input is not
required as Slack incoming webhooks are assigned to a channel.

An optional input accepts a map of alertmanager routes & receivers
along with Slack webhooks that are associated with the receivers.  Each
route is allocated a matcher of the same name so that alerts can be
routed to it.  These routes are in addition to the default routes.